pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.167.0",
    "google-auth>=2.39.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "routeros-api>=0.21.0",
]
